move ManifestViewUtils to base-tool
