+++ ui/base-tool/src/main/java/org/cru/godtools/base/tool/model/view/ManifestViewUtils.java
@@ -1,11 +1,10 @@
-package org.cru.godtools.tract.viewmodel;
+package org.cru.godtools.base.tool.model.view;
 
 import android.content.Context;
 import android.graphics.Typeface;
 import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
 
-import org.cru.godtools.base.tool.model.view.ResourceViewUtils;
 import org.cru.godtools.base.tool.widget.ScaledPicassoImageView;
 import org.cru.godtools.base.ui.util.LocaleTypefaceUtils;
 import org.cru.godtools.xml.model.Manifest;
